requirements.txt
tensorflow==2.12.0
pandas 
gdown
dvc
mlflow==2.2.2
# notebook==6.4.5      # pinned for Python 3.8 compatibility
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .